third_party/blink/web_tests/fast/beacon/beacon-basic.html

<!DOCTYPE HTML>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/sab-polyfill.js"></script>
<script>
description("Exercising the Beacon API");

shouldBeTrue("Object.getPrototypeOf(navigator).hasOwnProperty('sendBeacon')");
shouldBeEqualToString("typeof navigator.sendBeacon", "function");
shouldThrow("navigator.sendBeacon()");
shouldThrow("navigator.sendBeacon('http:')");
shouldThrow("navigator.sendBeacon('javascript:alert(1);')");
shouldThrow("navigator.sendBeacon('https:', new Uint8Array(new SharedArrayBuffer(10)))");

</script>
