third_party/WebKit/LayoutTests/platform/android/fast/beacon/beacon-basic-expected.txt

Exercising the Beacon API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Object.getPrototypeOf(navigator).hasOwnProperty('sendBeacon') is true
PASS typeof navigator.sendBeacon is "function"
PASS navigator.sendBeacon() threw exception TypeError: Failed to execute 'sendBeacon' on 'Navigator': 1 argument required, but only 0 present..
FAIL navigator.sendBeacon('http:') should throw an exception. Was true.
PASS navigator.sendBeacon('javascript:alert(1);') threw exception TypeError: Failed to execute 'sendBeacon' on 'Navigator': Beacons are only supported over HTTP(S)..
PASS successfullyParsed is true

TEST COMPLETE
